+++ test/unit-test/src/cf-expandables/src/Expandable-spec.js
@@ -4,7 +4,7 @@ const Expandable = require(
 );
 
 const HTML_SNIPPET = `
-<div class="o-expandable-group o-expandable-group__accordion"
+<div class="o-expandable-group"
      id="test-group-one">
 
     <div class="o-expandable o-expandable__padded" id="test-subject-one">
@@ -63,12 +63,15 @@ const HTML_SNIPPET = `
 
 let expandableDom1;
 let expandableDom2;
+let _expandables;
+let _expandable;
 
 describe( 'Expandable', () => {
 
   beforeEach( () => {
     document.body.innerHTML = HTML_SNIPPET;
-    Expandable.init();
+    _expandables = Expandable.init();
+    _expandable = _expandables[0];
     expandableDom1 = document.querySelector( '#test-subject-one' );
     expandableDom2 = document.querySelector( '#test-subject-two' );
   } );
@@ -81,7 +84,7 @@ describe( 'Expandable', () => {
         '.o-expandable_content'
       );
       expect( contentDom.offsetHeight ).toBe( 0 );
-      expect( contentDom.style.maxHeight ).toBe( '' );
+      expect( contentDom.style.maxHeight ).toBe( '0' );
       expect( contentDom.classList.contains(
         'o-expandable_content__expanded'
       ) ).toBe( false );
@@ -93,11 +96,32 @@ describe( 'Expandable', () => {
 
   describe( 'interactions', () => {
     it( 'should expand on click', () => {
-      const targetDom = expandableDom1.querySelector( '.o-expandable_target' );
-      simulateEvent( 'click', targetDom );
+      /*
+        TODO: Test expandEnd after ExpandableTransition is moved into
+        cf-expandables
+      */
+      let expandBeginFired = false;
+      let transitionEndFired = false;
+
+      const targetDom = expandableDom1.querySelector(
+        '.o-expandable_target'
+      );
       const contentDom = expandableDom1.querySelector(
         '.o-expandable_content'
       );
+
+      _expandable.transition.addEventListener( 'expandBegin', () => {
+        expandBeginFired = true;
+      } );
+      _expandable.transition.addEventListener( 'transitionEnd', () => {
+        transitionEndFired = true;
+      } );
+
+      simulateEvent( 'click', targetDom );
+      simulateEvent( 'transitionEnd', contentDom );
+
+      expect( expandBeginFired ).toBe( true );
+      expect( transitionEndFired ).toBe( true );
       expect( contentDom.style.maxHeight ).not.toBe( '' );
       expect( contentDom.classList.contains(
         'o-expandable_content__expanded'
@@ -108,12 +132,32 @@ describe( 'Expandable', () => {
     } );
 
     it( 'should go back to initial state on second click', () => {
-      const targetDom = expandableDom1.querySelector( '.o-expandable_target' );
-      simulateEvent( 'click', targetDom );
-      simulateEvent( 'click', targetDom );
+      /*
+        TODO: Test collapseEnd after ExpandableTransition is moved into
+        cf-expandables
+      */
+      let collapseBeginFired = false;
+      let transitionEndFired = false;
+
+      const targetDom = expandableDom1.querySelector(
+        '.o-expandable_target'
+      );
       const contentDom = expandableDom1.querySelector(
         '.o-expandable_content'
       );
+
+      _expandable.transition.addEventListener( 'collapseBegin', () => {
+        collapseBeginFired = true;
+      } );
+      _expandable.transition.addEventListener( 'transitionEnd', () => {
+        transitionEndFired = true;
+      } );
+
+      simulateEvent( 'click', targetDom );
+      simulateEvent( 'click', targetDom );
+
+      expect( collapseBeginFired ).toBe( true );
+      expect( transitionEndFired ).toBe( true );
       expect( contentDom.offsetHeight ).toBe( 0 );
       expect( contentDom.style.maxHeight ).toBe( '0' );
       expect( contentDom.classList.contains(
@@ -124,7 +168,7 @@ describe( 'Expandable', () => {
       ) ).toBe( true );
     } );
 
-    it( 'should swap open expandable on accordion click', () => {
+    xit( 'should swap open expandable on accordion click', () => {
       const targetDom1 = expandableDom1.querySelector( '.o-expandable_target' );
       const targetDom2 = expandableDom2.querySelector( '.o-expandable_target' );
       simulateEvent( 'click', targetDom1 );